requirements.txt
python-jenkins
PyGithub
aiohttp
motor
pyyaml
pytest
pytest-asyncio
pytest-cov
pytest-mock
asynctest
mockito
pytest-mockito